problema ao renderizar categorias
